Supprt minisnip
Needs https://github.com/joereynolds/vim-minisnip/pull/5

I've also changed the default for `g:go_snippet_engine` to `automatic`,
which is a new value that will use the first snippet engine that's
loaded. I think that's more user-friendly than expecting the user to set
it.

I also changed the detection method to use the variables, instead of
scanning `runtimepath`. This is more efficient.

Need to actually add the snippets for minisnip though.