Use :sbuffer in :GoAlternate
Use the `:sbuffer` command when it's already loaded. This is especially
useful if switchbuf is set to useopen or usetab.
Otherwise setting g:go_alternate_mode to `split` or `tabe` will just
keep on creating new windows or tabs.

This also fixes a bug where paths with e.g. a space didn't work
correctly.

Also add a test for go#alternate#Switch() and a few help for tests. If
`g:go_running_tests` is set, `go#util#Echo*` will record messages in
`g:go_messages` so we can easily access it from tests.