Active buffer relative to workdir for gometalint
When go#lint#Gometa is triggered by autosave, we attempt to only show
lint warnings and errors that are applicable for the active buffer. This
is done by passing the path to the file represented by the active buffer
to gometalinter's `--include` option.

The existing behaviour expands the absolute path, whereas gometalinter
outputs paths that are relative to the current working directory.

This change modifies the variable used for matching so that we also use
the current working directory for matching against lint errors/warnings.

Signed-off-by: Tom Barlow <tomwbarlow@gmail.com>