separate gometalinter error list use cases
gometalinter is used in two distinct cases. When run using
:GoMetaLinter, all of its output is used. When run automatically on save
(e.g. g:go_metalinter_autosave=1), only its output that applies to the
current buffer is used. In that case, it makes sense to show its
messages in the location list without forcing the user to also use the
location list when the user executes :GoMetaLinter.

Therefore, add a new key to g:go_list_type_commands,
GoMetaLinterAutoSave, with a default value of locationlist and modify
go#lint#Gometa to distinguish between the two cases.

Fixes #1644