fix races in go#job#Spawn callbacks
There may still be data data in the channel when the job exits.
Conversely, the exit callback may be called after the channel is closed.
To handle both situations, only call callbacks with exit status and all
the messages once the job has exited and the channel has been closed.

Replace the error_info_cb and custom_cb keys with a single new key,
complete.