Only return callbacks from go#job#Spawn
Only return the necessary callbacks from go#job#Spawn. Data that's
needed during the execution of the callbacks is wrapped in a closure
instead.

Do not allow the caller to provide an exit_cb or close_cb; there are no
current use cases that do so, and further changes would need to be made
to ensure that the complete handler would be called if only one of
exit_cb or close_cb were called.