def: implement lazy evaluation of definition call
This speeds up guru/godef jumping immensely by computing the destination
before the user decided to jump. This makes it feel like any jump is
instantaneous as the evaluating is done in background.

Because we don't want to block the UI during the computing it's only
supported for async mode.