avoid race conditions in lint async jobs
Avoid race conditions in lint async jobs by using Partials instead of
closures so that all state is captured and remains unchanged for each
invocation. This is especially a concern long running linting jobs
triggered by saving when the user saves multiple or saves and then moves
to the next buffer and saves it quickly, too.