list: make Clean() more useful
After cleaning a list, there is no need to show the list anymore. That's
why we call `Window()` to invoke the close function. Remove all
`Window()` calls and make Clean() more useful to avoid adding the second
line. There is no single `Clean()` call independently, so it makes sense
to change the meaning