Add :GoDebugTest
Imagine I'm editing `a_test.go`, which is failing. To investigate I want
to add a breakpoint in a function in `a.go`. The current procedure for
that is something like:

- `:e a.go`
- `:GoDebugBreakpoint`
- `:e a_test.go`
- `:GoDebugStart`
- `<F5>`
- Buffer will switch back to `a.go`

The `:e_test.go` is needed because the only way to run `dlv test`
instead of `dlv debug` is to have the current buffer end in `_test.go`.

I think this is rather awkward, so add `:GoDebugTest` instead of trying
to guess what the users wants so you now can do:

- `:e a.go`
- `:GoDebugBreakpoint`
- `:GoDebugTest`
- `<F5>`

Fixes the remainder of #1690