Set appropriate initial breakpoint
The debugger will set a breakpoint on `main.main` when starting a
program without any breakpoints set. This doesn't work in tests.

With this change it will try to get the function the cursor is in and
break on that. Come to think of it, this behaviour perhaps also makes
sense for non-test behaviour? Not sure.

I split up `go#textobj#FunctionJump()` to get the function name, there
are no functional changes there except replacing `eval(out)` with
`json_decode(out)`.

Fixes #1729