Add :GoDebugTest (#1728)
* cd in to directory before debugging test files

This is also what "go test" seems to do:

    [~/go/src/a]% cat *.go
    package a

    import (
            "fmt"
            "os"
    )

    func a() {
            d, _ := os.Getwd()
            fmt.Println("a.go:", d)
    }
    package a

    import (
            "fmt"
            "os"
            "testing"
    )

    func TestA(t *testing.T) {
            a()
            d, _ := os.Getwd()
            fmt.Println("a_test.go:", d)
    }

    [~/go/src/a]% go test ./... -v
    === RUN   TestA
    a.go: /home/martin/go/src/a
    a_test.go: /home/martin/go/src/a
    --- PASS: TestA (0.00s)
    PASS
    ok      a       0.001s
    === RUN   TestB
    b_test.go: /home/martin/go/src/a/b
    b_test.go: /home/martin/go/src/a/b
    --- PASS: TestB (0.00s)
    PASS
    ok      a/b     0.001s

Seems to make sense to match this behaviour.

Fixes part of #1690

* Add :GoDebugTest

Imagine I'm editing `a_test.go`, which is failing. To investigate I want
to add a breakpoint in a function in `a.go`. The current procedure for
that is something like:

- `:e a.go`
- `:GoDebugBreakpoint`
- `:e a_test.go`
- `:GoDebugStart`
- `<F5>`
- Buffer will switch back to `a.go`

The `:e_test.go` is needed because the only way to run `dlv test`
instead of `dlv debug` is to have the current buffer end in `_test.go`.

I think this is rather awkward, so add `:GoDebugTest` instead of trying
to guess what the users wants so you now can do:

- `:e a.go`
- `:GoDebugBreakpoint`
- `:GoDebugTest`
- `<F5>`

Fixes the remainder of #1690

* Set appropriate initial breakpoint

The debugger will set a breakpoint on `main.main` when starting a
program without any breakpoints set. This doesn't work in tests.

With this change it will try to get the function the cursor is in and
break on that. Come to think of it, this behaviour perhaps also makes
sense for non-test behaviour? Not sure.

I split up `go#textobj#FunctionJump()` to get the function name, there
are no functional changes there except replacing `eval(out)` with
`json_decode(out)`.

Fixes #1729

* Actually, it's much simpler to just add a breakpoint on the current line