Show debug errors as errors and not progress
Show stderr as error, not progress.

It *still* only shows the last time, but that's the nature of outputting
stuff in a job callback: you don't get the "hit ENTER" prompt (which in
this case is *wanted*). I don't know if there is any way to fix that
(it's a problem we have in several commands)?

We can probably populate the quickfix like we do with `:GoInstall` and
such, but that's a PR for another day.

See #1742