Warn when running go#cmd#autowrite() without autowrite enabled (#1754)
The `go#cmd#autowrite()` will write buffers to disk for commands that
operate on files (rather than buffers). This is useful, but if you
*don't* have this option enabled you never get an error message.

This changed that to show a non-fatal warning. This is the same
behaviour as the `:make` command.

Right now this doesn't work very well because the various status
messages from the jobs push the error away, making it hard to see (I
don't get the expected "hit enter" prompt), so we need a better solution
for that first.