Allow re-running the last matched function with :GoTestFunc last
I'm not sure if `last` is the best keyword, but I think this is a nice
feature to add.

Also reword the documentation for `:GoTest` a bit.

As a related change, this drops `-o` from `:GoTestCompile`. The output
file was in the Vim instances tmpdir, which is in the form of
`/tmp/$random/$n.vim-go.test`. There is no easy way to figure out the
value of either `$random` nor `$n`, so actually getting to the binary is
hard (you can use `g:go_debug = ['shell-command']` that was added just a
few days ago).

So, I don't think anyone is using that output file, and it's just taking
up disk space in `/tmp` (since Go binaries tend to be large it can add
up).

Fixes #1780