Respect user debug highlight group customization
As mentioned in #1839, If the user sets GoDebugCurrent or
GoDebugBreakpoint, we should respect their customization and not
override them.

The problem is that we're defining it in autoload, so it gets evaluated
after user customizations have been applied. I suspect the right thing
to do is to move the definition of the groups & signs to the syntax
file, but I'm going to assume things were done the way they were for a
good reason, and not propose that change (I'm no vim-expert).

It's not obvious how we can check for this, so I've chosen to execute
"hi GoDebugBreakpoint" and look for a string that appears to be present
when it was not already set.