fix guru output handling in Neovim
Output from a Neovim job may come in batches and splits on newlines.
When the output ends with a newline, the last element in the data passed
to the stdout and stderr callbacks will be the empty string. When it's
not the empty string, then last element is an incomplete line, and the
next call will contain the rest of the line in the first element of its
data.

Fixes #1836