refactor vim8 job related functions
Rename go#job#Spawn to go#job#Options.

Add a new key, cwd, to the return value of go#job#Options so that the
working directory for the job can be passed in the job options.

Add a new function, go#job#Start, to wrap vim8's job_start function.
Eventually, this new function will be used to start Neovim jobs, too, so
that vim-go can have an abstraction to start jobs that hides the details
of dealing with Neovim and Vim8 job API differences.