restore behavior of :GoInfo when g:go_info_mode='gocode'
Make sure :GoInfo echoes the info for the identifier under the cursor
when g:go_info_mode='gocode'.

Remove the auto argument from go#complete#Info; it is not needed. In
other locations, rename auto to echo so that it is clear what effect it
has instead of acting as a signal for _how_ it was called.

Fixes #1914.