Shutdown and detach from delve correctly
Use a single connection to delve to simplify detaching properly. The
previous implementation was making a connection to delve for each
request and was never closing any of them. Instead, create a single
connection to delve and reuse it for each rpc call. This allows vim-go
to not have to pass `--accept-multiclient`, which in turn shutsdown
delve when detach is called.

As part of this work, `call_jsonrpc` was refactored to not deal with
callbacks explicitly. Because Neovim requires a callback to get data
from a socket, the `on_data` callback was moved into `out_cb` and
improved to deal with the edge cases that Neovim has with regard to
partial messages.