clear buffer-local autocmds correctly
Clear buffer-local autocmds correctly without clearing the same autocmds
for other buffers.

As an example of why this should be done: prior to this change, opening
a window for each of two separate files and then running `:GoCoverage`
in each would highlight coverage as expected. Loading a third buffer
into one of the windows would clear the highlighting in that window and
would _also_ remove the other window's buffer-local autocmd event to
clear its coverage highlighting. Loading a different buffer into the
other window would unexpectedly _keep_ the highlighting, because the
autocmd event configuration was removed when the third buffer was loaded
into its window.