Prefer the system PATH for some commands
The default $PATH is a better place to look than $GOPATH/bin for some
tools. This is especially true for 'go', 'diff', and 'uname'.

If someone incidentally creates a go binary named 'go' or the like, but
has configured their $PATH to prefer the system 'go' binary to the one
in their $GOPATH, it seems unexpected to go ahead and run it anyway.

For the tools that are expected to be in the $GOPATH/bin folder, like
'goimports', it does make sense to just go ahead and check that folder
first.