Use visualfc/gocode
There are multiple versions of gocode now - the original nsf/gocode but
it's deprecated in favor of mdempsky/gocode. The latter is not
maintained anymore and doesn't support modules so there are multiple
versions of gocode with support for modules - stamblerre/gocode and
visualfc/gocode. The stamblerre was used as a separate binary
gocode-gomod alas it's not working all the time (see for example end of
the discussion https://github.com/Microsoft/vscode-go/pull/1932).

Meanwhile, visualfc/gocode is working great both for GOPATH code and modules
code so there is no need to support 2 binaries.

In these changes we switch to visualfc/gocode, remove stamblerre/gocode
version and also remove unnecessary choise of binary.

Fixes #1906.