go/lsp: don't collect when output is of type string
`a:msg` is of type string if gopls can't resolve a completion. This is
reproducible if you write a random identifier and try to complete (I
discovered when I mispelled a variable). Such as `reps.  <<<< complete here`

However this is not working well, somehow vim-go is blocking and I have
to press C-c. I think we're sleeping or blocking somewhere but couldn't
find the place for it. This is not mergable therefore but opening in
case you know how to fix it properly.