synchronize buffers with gopls
Notify gopls when buffers have changed. This won't catch all use cases,
but it least makes sure that gopls is notified when the the file is
changed on disk (e.g. via `:GoFmt`, `:GoRename`, or other actions by the
user) and when gopls should consider the file to be closed, because the
buffer was deleted.