debug: allow debugging outside GOPATH without a go.mod
Do not try to determine the package name being debugged. Delve supports
debugging and testing without specifying a package name; vim-go should,
too. One of the things that this makes easier is debugging code that is
neither in GOPATH nor in a module.

Remove unnecessary is_test key from the state dictionary.

Remove code that set the current directory when debugging tests. Because
delve is started as an async job, its cwd will be set to the directory
of the buffer from which debugging is started, anyway. Using `lcd` to
set the buffer local directory is just another thing to keep track of
and it doesn't seem to have any real benefit.

Update documentation to explain how relative paths work for `:GoDebug`
and `:GoDebugTest`.

Refactor go#package#FromPath so that it returns -2 when the package is
neither in GOPATH nor in the current module.