test: provide output when examples fail
Output the name of the example when an example tests fails. Example
failures do not provide a file or line number, and the output isn't
indented, so it's not easy to show the full output, but by providing at
least the name in the quickfix list the user will know what failed.

Fixes #2271