package/FromPath: Use mod path if there's no code
To determine the import path of a directory, vim-go currently relies on
`go list`.

With Go modules, this has a few side effects, one of which is that it
makes a number of HTTP requests if there is no code in the current
directory. For example,

    $ export GO111MODULE=on
    $ cd $GOPATH/src
    $ mkdir my.example/foo/bar/baz/qux
    $ cd my.example/foo/bar/baz/qux

    $ go mod init
    go: creating new go.mod: module my.example/foo/bar/baz/qux

    $ go list -v
    Fetching https://my.example/foo/bar/baz?go-get=1
    https fetch failed: Get https://my.example/foo/bar/baz?go-get=1: dial tcp: lookup my.example: no such host
    Fetching https://my.example/foo/bar?go-get=1
    https fetch failed: Get https://my.example/foo/bar?go-get=1: dial tcp: lookup my.example: no such host
    Fetching https://my.example/foo?go-get=1
    https fetch failed: Get https://my.example/foo?go-get=1: dial tcp: lookup my.example: no such host
    Fetching https://my.example?go-get=1
    https fetch failed: Get https://my.example?go-get=1: dial tcp: lookup my.example: no such host
    can't load package: package my.example/foo/bar/baz/qux: unknown import path "my.example/foo/bar/baz/qux": cannot find module providing package my.example/foo/bar/baz/qux

Whether this behavior is desirable is still under discussion (see
golang/go#29452), but meanwhile this is affecting the experience of
using vim-go since bfe04ba.

Specifically, if the domain (`my.example`) points to an address that
takes some time to respond, it takes quite a while for vim to be ready
upon opening a new file. For example, if the timeout is even 3 seconds,
it gets multiplied by the number of components in the import path.

As a workaround for this, this changes FromPath to check if there is Go
code in the current directory before running `go list`.

If there is no Go code in the current directory, we'll build the import
path for it by combining the module path with the relative path to the
current directory.

So for the example above,

    $ cd my.example/foo/bar/baz/qux
    $ vim +"echo go#package#FromPath(expand('%'))" test.go
    # ==> my.example/foo/bar/baz/qux

    $ mkdir -p cmd/
    $ vim +"echo go#package#FromPath(expand('%'))" cmd/main.go
    # ==> my.example/foo/bar/baz/qux/cmd