lsp: refactor how g:go_lsp_log is represented
Change g:go_lsp_log from an array of strings to an array of
dictionaries. Each dictionary consists of a single item whose key is
either request, response, or stderr to represent whether the item was a
message sent to gopls, a response received from gopls, or stderr from
gopls, respectively.