try to use relative import paths if necessary
Teach go#debug#Start, go#lint#Vet, and go#lint#ErrCheck to work in null
modules.

Refactor go#package#ImportPath to delegate to go#package#FromPath.

Teach go#package#FromPath to return a relative package path when in a
null module if possible. When the package is not within the current
working directory, -2 will still be returned. This allows
go#debug#Start, go#lint#Vet, and go#lint#Errcheck to work in null
modules.

Add a test for go#debug#Start using a relative path in a null module.

Correct documentation for :GoDebugStart.

Fix `:GoVet` with arguments so that it will run for the package that
contains the current buffer instead of just using whatever the current
working directory is.