lsp: make initialization of gopls in uninitialized modules more responsive
When go list is run in a module that has uncached requirements, go list
will reach out to the internet to get those requirements. Because gopls
doesn't work in a null module, vim-go has to run go list to determine
whether a buffer is in a null module or not when deciding whether to
initialize gopls. When the buffer is in a non-null module, go list can
cause a delay. To make Vim response while go list is running, determine
whether a module is a null module in a timer and sleep in a loop while
waiting for the timer callback to complete. The sleep while waiting on
the callback to complete allows Vim to be responsive while deciding
whether to initialize gopls.

Remove some TODO statements that identify tasks that do not need to be
done

Simplify checking dictionary key values in some locations to improve
readability..