plugin: install gopls stable
Refactor plugin/go.vim to support installing gopls' stable releases by
adding support for specifying which module version to install. When the
package name includes a `@`, module mode will be used.

Set and restore GO111MODULE for _each_ binary to be installed instead of
setting it once for _all_ binaries to be installed.