lsp: handle changes to function and method completions
gopls is changing the values in the label and details of completions so
that the label is the name of the function or method and the detail is
the full signature w/ return value. Handle both the new way and the old
way for backward compatibility until golang.org/x/tools/gopls@latest
serves the new way and users have had a reasonable time to update.