lsp: fix off by one in go#complete#Complete
Now that go#complete#Complete handles LSP's expression of the position
as utf-16 code units, return the start position correctly. The start
position provided by LSP is the index of the first code unit that starts
the completion text, while Vim expect to be provided with the number of
leading columns before the start of the text to be replaced with the
completion item. Therefore, go#lsp#lsp#PositionOf is given the LSP start
position - 1 code unit so that all the characters prior to the start of
the completion text will left in place by Vim.

Fixes #2403