snippets: revert #2358
There were two of them, one is `errn` and errn,` (note the coma at the
end). We didn't use `err` because that is sometimes a variable itself
and you don't want to autoexpand it. Or sometimes the completer catches
it.

We should revert it because first because `errn` is not in use at all,
so this change was wrong in the first pleace and this breaks things for
many people who were using `errn` before. It's not a backwars compatible
change.