debug: refactor goroutine UX
Refactor debug window placement so that the windows are sized and place
properly simply by creating the splits so that windows don't have to be
resized after the splits are created.

Factor the showing of goroutines into its own method for readability and
consistency with how other windows are handled.

Return early without trying to change the goroutine when the user
presses enter on a line that doesn't list a goroutine or when the chosen
goroutine is already current.

Use window management functions instead of normal mode commands.

Use 'ID' consistently instead of 'Id'.