def: wrap switching buffers in a try/catch
Wrap the loading of the target buffer in a try/catch so that the
function doesn't abort if the buffer is loaded in another Vim session.

Unforunately, there is a problem with this solution, too: the
buffer ends up being empty, and one has to execute `:edit` to reload it
to get the file to display in the window.

Fixes #2504.