plugin: handle build flags in module aware mode
Handle build flags when installing binaries in module aware mode.

Make the default value for builds flags an empty array instead of a
string for consistency.

Set build flags before the output path and the package name in GOPATH
mode to be consistent with the order in the module aware mode, too.