resolve :GoReferrers errors
Fix errors that occurred when running :GoReferrers more than once when
one of the references is in a file that is not loaded. On the first run,
readline() would be used to get the contents, but then on the second run
there is a buffer for the file that was previously not assigned a buffer
number, but it's an unloaded buffer, so getbufline() returned an empty
list.