lsp: handle completions with placeholders as snippets
Handle completions from gopls with placeholders as anonymous snippets
with UltiSnips.

Change go#config#SnippetEngine to return the snippet engine to be used
instead of just the raw value of g:go_snippet_engine so that it can be
more effectively reused instead of interpreting the return value in both
ftplugin/go/snippets.vim and autoload/go/auto.vim.