clear matches when a buffer is removed from a window
Clear diagnostic and sameid matches when a new buffer is loaded into a
window. The matches are not yet cleared when the previous buffer is also
loaded into another window; I'm looking for a clean way to do that that
isn't too heavy, but I haven't found it yet.

Highlight diagnostic errors and warnings when an existing buffer whose
errors and warnings are known is loaded into a window.

Work around matchaddpos() limitation of only supporting 8 positions per
call by introducing a new wrapper function, go#util#MatchAddPos(), that
will make multiple calls to matchaddpos().

Fixes #2631