define text property types in syntax/go.vim
Define text property types in syntax/go.vim so that they won't be added
until the syntax file is loaded; syntax files will be loaded after
ftplugins when `syntax on` or `syntax enable` is called after `filetype
plugin on`.

Do not try to remove a text property whose type does not exist in order
to avoid errors when syntax highlighting is not enabled.

Fixes #2658