term: fix terminal window sizing
Split the terminal for Vim the same as is done for Neovim and then start
the terminal in that new buffer.

Fix a regression in Neovim's terminal window that caused it to be too
small because the height and width were based on the starting window
instead of terminal window split.

Set the window to the starting window on exit.

Fixes #2735