lsp: drop out of date diagnostics
Drop out of date diagnostics so that highlighting will not be applied
incorrectly.

Running :GoImports was causing errors sometimes and
causing highlighting to be applied on the wrong lines since it changed
the source. Since formatting with gofmt or goimports will change the
files, clear the diagnostic errors and warnings before reloading the
buffer so that the BufWinEnter autocmd doesn't try to set the matches to
the wrong locations.

Send changes to gopls after formatting with gofmt or goimports so that
gopls will have the current version.