lsp: drop out of date diagnostics
Drop out of date diagnostics so that highlighting will not be applied
incorrectly. Running `:GoImports` was causing errors sometimes and
causing highlighting to be applied on the wrong lines since it changed
the source.

Send changes to gopls after formatting with gofmt or goimports so that
gopls will have the current version.