run: capture error messages from panics
Capture error messages from panic. I'm not 100% sure that this is the
right course of action. It might be better to treat panics as successes,
because non-zero exit status from go run are generally discarded except
for compiler errors. This change is inconsistent with that fact and may
lead people to expect that non-zero exit statuses from the command being
run should always populate the quickfix list.

Fixes #2328