Add support for running a single inner test.
This allows GoTestFunc to run a single t.Run() inner test rather than
running all of the inner tests within a single test function. It will
only run an inner test if the user's cursor is inside the test
declaration. If the cursor is on the test definition itself then it will
run the entire test. If the cursor is within the definition it will run
the first inner test that is on a line before or equal to the current
cursor line. If there are no inner tests then the entire test method
will be run normally.